+++ src/commonMain/kotlin/space/kscience/gdml/solids.kt
@@ -308,12 +308,14 @@ public sealed class GdmlBoolSolid : GdmlSolid() {
     public fun resolveFirstPosition(root: Gdml): GdmlPosition? = firstposition ?: firstpositionref?.resolve(root)
     public fun resolveFirstRotation(root: Gdml): GdmlRotation? = firstrotation ?: firstrotationref?.resolve(root)
 
+    @Deprecated("Use GdmlPosition constructor instead")
     public fun position(x: Number = 0f, y: Number = 0f, z: Number = 0f): GdmlPosition = GdmlPosition().apply {
         this.x = x
         this.y = y
         this.z = z
     }
 
+    @Deprecated("Use GdmlRotation constructor instead")
     public fun rotation(x: Number = 0f, y: Number = 0f, z: Number = 0f): GdmlRotation = GdmlRotation().apply {
         this.x = x
         this.y = y